backend/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
elasticsearch==8.11.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
